+++ R/check_config.R
@@ -16,38 +16,35 @@ check_config = function() {
   front = stringr::str_split(front, "\\\\")[[1]][1]
   front = stringr::str_trim(front, side = "right")
   if (stringr::str_length(front) > 25L) {
-    msg_error("Title is too long. Add a line break?", padding = 2)
-    set_error()
+    msg_error("Title is too long. Add a line break?", padding = TRUE)
     config_issue = TRUE
   }
 
   if (is.null(config::get("advert"))) {
-    msg_error("Advert missing from config.", padding = 2)
-    msg_error("Add 'advert: advert'", padding = 4)
-    set_error()
+    msg_error("Advert missing from config.", padding = TRUE)
+    msg_error("Add 'advert: advert'", padding = TRUE)
     config_issue = TRUE
   }
 
   if (is.null(config::get("courses"))) {
-    msg_info("Course dependency graph has not been included.", padding = 2)
-    msg_info("Add 'courses: course-dependencies' to the config.yml", padding = 4)
+    msg_info("Course dependency graph has not been included.", padding = TRUE)
+    msg_info("Add 'courses: course-dependencies' to the config.yml", padding = TRUE)
     config_issue = TRUE
   }
 
   p = config::get("packages")
   if (!is.null(p)) {
     lang = get_repo_language()
-    msg_error("Old style pkg config detected. Please remove", padding = 2)
-    msg_error("Instead use:", padding = 2)
+    msg_error("Old style pkg config detected. Please remove", padding = TRUE)
+    msg_error("Instead use:", padding = TRUE)
     message(glue::glue("  {lang}_packages:"))
     message("    - package1")
     message("    - package2")
-    set_error()
     config_issue = TRUE
   }
 
   if (isFALSE(config_issue)) {
-    msg_ok("Config looks good!")
+    msg_success("Config looks good!")
   }
   return(invisible(NULL))
 }