+++ R/tokens.R
@@ -1,14 +1,14 @@
 tokenise = function() {
-  fname = system.file("extdata", "extractor.tex",
+  msg_start("Creating tokens...tokenise()")
+    fname = system.file("extdata", "extractor.tex",
                       package = "jrNotes", mustWork = TRUE)
 
   file.copy(fname, to = "extractor.tex", overwrite = TRUE)
   file.copy("main.tex", "extractor-tmp.tex", overwrite = TRUE)
-  msg_info("Checking tokens", padding = 2)
   system2("xelatex",
           args = c("'\\input extractor \\input extractor-tmp.tex'"),
           stdout = FALSE)
-  msg_ok("Tokens created", padding = 2)
+  msg_success("Tokens created")
 }
 
 globalVariables("text")